Rename anomalib.deploy.inferencer.*_inference.py to inferencer.py
