+++ anomalib/deploy/inferencers/torch_inferencer.py
@@ -29,7 +29,7 @@ from anomalib.models import get_model
 from anomalib.models.components import AnomalyModule
 from anomalib.pre_processing import PreProcessor
 
-from .base_inference import Inferencer
+from .base_inferencer import Inferencer
 
 
 class TorchInferencer(Inferencer):
